+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Unreleased
+- Include pyenv and Python 3.9 to 3.12 in the kitchen sink image
+
+  ([#232](https://github.com/pulumi/pulumi-docker-containers/pull/232))
 
 - Add $GOPATH/bin to $PATH for Go containers
   ([249](https://github.com/pulumi/pulumi-docker-containers/pull/249))